Add createTestUser function and refactor student user generation logic
